feat: Add mutliple API calling function
Use Observable to handle mutliple API calling.
Install RxJava2 to import the Observable library.
Use the Zip function to make multiple requests.

All requests are successfully made.  Left with
handling the Callback.

The existing ApiSerice made use of Retrofit.Call
to trigger the Callback function. It needs to
find a way to trigger the function in Zip function.